SysCallEnv.cpp
Program that allows the user to write their requested environment variables to a file. This program writes the requested environment variables' value into a given file. 

given in the format of " ./a.out FILENAME ENV ", FILENAME refers to the file which the user wishes to store the value for the requested environment variable, and ENV refers to the environment variable whose value the user wishes to see.